Prefer `Gem::Version` to `Chef::Version` when comparing versions
`Chef::Version` is compliant with Chef versioning which allows only
strict `major.minor.patch` where major/minor/patch are integer. For
example it raise an error with version such as `1.0.1b`, `1.0.1.1` or
`1.0.0.beta1`. As we're parsing external software versions it's better
to be more permissive, thus using `Gem::Version`, which works for all
kind of version string.

Rubygems being included in Ruby 1.9+, and `Gem::Version` being available
since RubyGems `1.4.0`, it should be safe to rely on it.